server/requirements.txt
flask==3.0.3
gunicorn==23.0.0
pandas==2.2.3
openai==1.53.0
python-dotenv==1.0.1
pandas==2.2.3
chardet==5.2.0